requirements.txt
openai
python-dotenv
langchain
langchain-openai
langchain-community
azure-search-documents
azure-identity
pypdf
pandas
streamlit